+++ src/main/java/codifica/eleve/core/domain/solicitacao/SolicitacaoAgenda.java
@@ -12,7 +12,7 @@ import java.util.List;
 public class SolicitacaoAgenda {
 
     private Id id;
-    private Integer chatId;
+    private Long chatId;
     private Pet pet;
     private List<Servico> servicos;
     private ValorMonetario valorDeslocamento;
@@ -21,7 +21,7 @@ public class SolicitacaoAgenda {
     private LocalDateTime dataHoraSolicitacao;
     private String status;
 
-    public SolicitacaoAgenda(Integer chatId, Pet pet, List<Servico> servicos, ValorMonetario valorDeslocamento, LocalDateTime dataHoraInicio, LocalDateTime dataHoraFim, LocalDateTime dataHoraSolicitacao, String status) {
+    public SolicitacaoAgenda(Long chatId, Pet pet, List<Servico> servicos, ValorMonetario valorDeslocamento, LocalDateTime dataHoraInicio, LocalDateTime dataHoraFim, LocalDateTime dataHoraSolicitacao, String status) {
         if (servicos == null || servicos.isEmpty()) {
             throw new IllegalArgumentException("A lista de serviços não pode ser vazia.");
         }
@@ -56,11 +56,11 @@ public class SolicitacaoAgenda {
         this.id = id;
     }
 
-    public Integer getChatId() {
+    public Long getChatId() {
         return chatId;
     }
 
-    public void setChatId(Integer chatId) {
+    public void setChatId(Long chatId) {
         this.chatId = chatId;
     }
